Remove ParticipantsMiddleware
