+++ AzureCommunicationUI/sdk/AzureCommunicationUIChat/Sources/DI/DependencyContainer.swift
@@ -81,13 +81,11 @@ final class DependencyContainer {
 
         let repositoryMiddlewareHandler = RepositoryMiddlewareHandler(messageRepository: resolve(),
                                                                           logger: resolve())
-        let participantsActionHandler = ParticipantsActionHandler(chatService: resolve(), logger: resolve())
         let middlewares: [Middleware] = [
             Middleware<AppState>.liveChatMiddleware(
                 chatActionHandler: chatActionHandler,
                 chatServiceEventHandler: chatServiceEventHandler),
-            Middleware<AppState>.liveRepositoryMiddleware(repositoryMiddlewareHandler: repositoryMiddlewareHandler),
-            Middleware<AppState>.liveParticipantsMiddleware(participantsActionHandler: participantsActionHandler)
+            Middleware<AppState>.liveRepositoryMiddleware(repositoryMiddlewareHandler: repositoryMiddlewareHandler)
         ]
 
         guard let localUserId = localUserIdentifier else {
